apps/api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
yt-dlp==2024.12.13
openai==1.58.1
ffmpeg-python==0.2.0
python-dotenv==1.0.1
python-multipart==0.0.20
aiofiles==24.1.0
